fix(overlay): pin thread toggle beside folder; hide chrome and corner mascot
Move the transcript button to the composer toolbar, left of the project folder. Collapsed overlay drops the title/close chrome so only the input and icon row remain. Hide the bottom-right mascot while the overlay is open so it cannot cover the composer.
