+++ packages/core-chat/src/web/composer-stack.test.ts
@@ -38,8 +38,12 @@ describe('composer dock stacking above sticky user', () => {
     const composer = readFileSync(resolve(root, 'packages/core-chat/src/web/composer.tsx'), 'utf8')
     expect(css).toMatch(/\.chat-overlay-panel \{/)
     expect(css).toMatch(/\.chat-overlay-panel\.is-compose-only \.chat-overlay-thread\s*\{[^}]*display:\s*none/s)
+    expect(css).toMatch(/\.chat-overlay-panel\.is-compose-only \.chat-overlay-head\s*\{[^}]*display:\s*none/s)
     expect(css).toMatch(/\.chat-overlay-panel\.is-compose-only\s*\{[^}]*background:\s*transparent/s)
     expect(css).toMatch(/\.chat-overlay-panel\.is-compose-only\s*\{[^}]*border:\s*0/s)
+    const overlayWin = readFileSync(resolve(root, 'packages/web-app-shell/src/web/overlay-window.tsx'), 'utf8')
+    expect(overlayWin).toMatch(/threadOpen \? \(/)
+    expect(overlayWin).toContain('chat-overlay-head')
     expect(css).not.toMatch(/chat-overlay-peek/)
     expect(composer).toContain('revealOverlayThread')
     expect(composer).toContain('isComposerFocusPending')
@@ -62,6 +66,7 @@ describe('composer dock stacking above sticky user', () => {
     expect(composer).not.toContain('CaretMascot')
     expect(composer).not.toContain('caret-mascot')
     expect(css).toMatch(/\.chat-dock-toolbar \.brand-corner-cluster\s*\{[^}]*position:\s*relative/s)
+    expect(approvals).toMatch(/chat-overlay-thread-toggle[\s\S]*SessionProjectPanel/)
     expect(approvals).not.toContain('dock-session-mascot')
     expect(approvals).not.toContain('BrandAgentMenu')
     expect(approvals).not.toContain('setSessionPickerOpen')
