+++ packages/web-app-shell/src/web/brand-corner.test.ts
@@ -49,6 +49,7 @@ test('brand mascot lives at the corner; sidebar head is title plus collapse', ()
   assert.doesNotMatch(mascot, /brand-corner-chat-overlay/)
   assert.doesNotMatch(mascot, /onToggleOverlay/)
   assert.match(shell, /ShellDockPins/)
+  assert.match(shell, /if \(overlayOpen\) return null/)
   assert.match(shell, /onToggle=\{/)
   assert.match(shell, /isChatPagePath\(location.pathname\)/)
   assert.match(shell, /setChatOverlay\(true\)/)
