+++ packages/web-app-shell/src/web/chat-decouple.test.ts
@@ -22,8 +22,10 @@ test('overlay chat is a resident floating window off the agent page', () => {
   assert.doesNotMatch(shell, /toggleChatOverlay/)
   assert.doesNotMatch(shell, /chat-overlay-pin/)
   const overlayHead = shell.slice(shell.indexOf('const overlayHeader'))
-  assert.match(overlayHead.slice(0, 2200), /chat-overlay-thread-toggle[\s\S]*FolderGlyph/)
+  assert.doesNotMatch(overlayHead.slice(0, 2200), /chat-overlay-thread-toggle/)
   assert.match(overlayHead.slice(0, 1800), /chat-view-project/)
+  const approvals = readFileSync(resolve(import.meta.dirname, '../../../core-chat/src/web/approvals.tsx'), 'utf8')
+  assert.match(approvals, /chat-overlay-thread-toggle[\s\S]*SessionProjectPanel/)
   assert.doesNotMatch(overlayHead.slice(0, 1600), /layoutTools/)
   assert.doesNotMatch(overlayHead.slice(0, 1600), /chat-overlay-layout/)
   assert.match(overlayHead.slice(0, 1600), /chat-overlay-close/)
