+++ packages/web-app-shell/src/web/index.tsx
@@ -5,9 +5,6 @@ import {
   subscribeChatOverlay,
   setChatOverlay,
   closeChatOverlay,
-  getOverlayThread,
-  subscribeOverlayThread,
-  toggleOverlayThread,
   requestOverlayFocus,
   requestInspectorClose,
   allocateShellColumns,
@@ -57,7 +54,6 @@ import {
   ChevronDoubleLeftIcon,
   ChevronDoubleRightIcon,
   AdjustmentsHorizontalIcon,
-  ChatBubbleLeftRightIcon,
   XMarkIcon,
 } from '@heroicons/react/16/solid'
 
@@ -104,6 +100,8 @@ function ShellDockPins({
 }: {
   useSessionView: ReturnType<typeof bindSessionView>
 }) {
+  const overlayOpen = useSyncExternalStore(subscribeChatOverlay, getChatOverlay, () => false)
+  if (overlayOpen) return null
   return <DockSessionMascot useSessionView={useSessionView} />
 }
 
@@ -235,7 +233,6 @@ function Shell(props: SlotProps) {
   const slots = props.slots as SlotsService
   const navigate = useNavigate()
   const location = useLocation()
-  const overlayThread = useSyncExternalStore(subscribeOverlayThread, getOverlayThread, () => false)
   const modules = useAppModules()
   const navReady = useAppModulesNavReady ? useAppModulesNavReady() : true
   const pluginModules = modules.filter((item) => item.id !== 'agent')
@@ -695,25 +692,12 @@ function Shell(props: SlotProps) {
   const overlayHeader = (
     <header className="chat-view-header" data-biu-ignore>
       <div className="chat-view-header-left">
-        <div className="chat-view-project" title={project ? (project.path ?? project.name) : undefined}>
-          <button
-            type="button"
-            className={`chat-view-header-expand${overlayThread ? ' is-active' : ''}`}
-            title={overlayThread ? '收起对话' : '查看对话'}
-            aria-label={overlayThread ? '收起对话' : '查看对话'}
-            aria-pressed={overlayThread}
-            data-testid="chat-overlay-thread-toggle"
-            onClick={() => toggleOverlayThread()}
-          >
-            <ChatBubbleLeftRightIcon {...chromeIcon} />
-          </button>
-          {project ? (
-            <>
-              <FolderGlyph className="chat-view-project-icon" />
-              <span className="chat-view-project-name">{project.name}</span>
-            </>
-          ) : null}
-        </div>
+        {project ? (
+          <div className="chat-view-project" title={project.path ?? project.name}>
+            <FolderGlyph className="chat-view-project-icon" />
+            <span className="chat-view-project-name">{project.name}</span>
+          </div>
+        ) : null}
       </div>
       <ChatSessionTitle useSessionView={useSessionView} sessionView={sessionView} />
       <div className="chat-view-header-right">
